% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_new_column}
\alias{add_new_column}
\title{Add a new column based on condition}
\usage{
add_new_column(df, col1_name, col2_name, new_col_name, other_list)
}
\description{
Add a new column based on condition
}
