% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout_validation.R
\name{clade_tips}
\alias{clade_tips}
\title{Tip labels below a given node}
\usage{
clade_tips(phy, node)
}
\arguments{
\item{phy}{Tree of class phylo.}

\item{node}{Node number.}
}
\value{
Character vector of tip labels.
}
\description{
Tip labels below a given node
}
