% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{combined_estimates}
\alias{combined_estimates}
\title{Collapse results files into a single dataframe with parameter estimates.}
\usage{
combined_estimates(x)
}
\description{
Collapse results files into a single dataframe with parameter estimates.
}
