% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout_alternating.R
\name{compute_intercepts}
\alias{compute_intercepts}
\title{compute_intercepts}
\usage{
compute_intercepts(betamat, meanx, sdx, meany, sdy)
}
\arguments{
\item{betamat}{a matrix where each column is a beta_hat estimate for some
regularization parameter}

\item{meanx}{a numeric vector. column means of original training X data}

\item{sdx}{a numeric vector. column SDs of original training X data}

\item{meany}{a numeric value. mean of original training Y data}

\item{sdy}{a numeric value. SD of original training Y data}
}
\description{
computes the intercept term for each beta_hat estimate.
The beta matrix should be such that each column contains a different beta_hat
}
