% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout_alternating.R
\name{compute_rmspe_betamat}
\alias{compute_rmspe_betamat}
\title{compute_rmspe_betamat}
\usage{
compute_rmspe_betamat(X_test, Y_test, intercepts, betamat)
}
\arguments{
\item{X_test}{n_test x p data matrix (test data, unstandardized)}

\item{Y_test}{n_test x 1 response vector (test data, unstandardized)}

\item{intercepts}{numeric vector of length nlambda}

\item{betamat}{numeric matrix of dimensions p x nlambda}
}
\description{
computes the rmspe between the provided Y_test and the predicted
X_test %*% beta_hat for each beta_hat in the provided beta matrix.
The beta matrix should be such that each column contains a different beta_hat
}
