% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{create_directory_if_not_exists}
\alias{create_directory_if_not_exists}
\title{Create directory if it does not exist}
\usage{
create_directory_if_not_exists(directory_path, message = TRUE)
}
\description{
Create directory if it does not exist
}
