% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout_alternating.R
\name{cv.scout_alternating_lasso}
\alias{cv.scout_alternating_lasso}
\title{cv.scout_alternating_lasso}
\usage{
cv.scout_alternating_lasso(
  X_train,
  Y_train,
  X_test,
  Y_test,
  p1,
  nlambda1 = 100,
  nlambda2 = 100,
  lambda1_min_ratio = 0.1,
  lambda2_min_ratio = 0.001,
  K = 5,
  tol = 1e-04,
  max_iter = 10,
  rescale = TRUE,
  standardize = TRUE,
  lam1_init = "random"
)
}
\description{
implements an alternating search algorithm to efficiently
search over a grid of lambda1 x lambda2. L_p penalty for coefficient
regularization is assumed to be 1 (i.e., p2 = 1). splits training data into
K folds and picks the result corresponding to fold with lowest RMSPE on
validation data.
}
