% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOUT_EM_utils.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Function to get parameters out of the model.}
\usage{
extract_parameters(results_list)
}
\description{
Function to get parameters out of the model.
}
