% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_thetas}
\alias{generate_thetas}
\title{Utility function to generate a vector of thetas.}
\usage{
generate_thetas(x, nstats, step)
}
\description{
Utility function to generate a vector of thetas.
}
