% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{get_fit_metrics}
\alias{get_fit_metrics}
\title{Calculate fit metrics and weights from model summaries}
\usage{
get_fit_metrics(results, outpath = NULL, prefix = "OUWIE", write = FALSE)
}
\arguments{
\item{outpath}{Directory to write weighted results CSV}

\item{prefix}{File prefix}

\item{write}{Logical, whether to write weighted results CSV}

\item{models_summary}{Data frame returned from fitModel}
}
\value{
Data frame with AICc ranks, delta and weights
}
\description{
Calculate fit metrics and weights from model summaries
}
