% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sheets.R
\name{get_ground_truth_cna}
\alias{get_ground_truth_cna}
\title{Get SCOUT ground truth CNA data}
\usage{
get_ground_truth_cna(spn = NULL, sample = NULL)
}
\arguments{
\item{spn}{SPN identifier, e.g. \code{"SPN01"}. If \code{NULL}, returns all SPNs.}

\item{sample}{Sample identifier. If \code{NULL}, returns all samples for the
selected SPN.}
}
\value{
A tibble.
}
\description{
Returns a tibble of ground truth copy number segments from the SCOUT
metadata Google Sheet, with columns \code{SPN}, \code{Sample}, \code{Chromosome},
\code{Start}, \code{End}, \code{Major}, \code{minor}, and \code{CCF}.
Optionally filtered to a specific SPN and/or sample.
}
\examples{
\dontrun{
get_ground_truth_cna()
get_ground_truth_cna("SPN01")
get_ground_truth_cna("SPN01", sample = "1.1")
}
}
