% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sheets.R
\name{get_ground_truth_drivers}
\alias{get_ground_truth_drivers}
\title{Get SCOUT ground truth driver events}
\usage{
get_ground_truth_drivers(spn = NULL)
}
\arguments{
\item{spn}{SPN identifier, e.g. \code{"SPN01"}. If \code{NULL}, returns all SPNs.}
}
\value{
A tibble.
}
\description{
Returns a tibble of ground truth driver events from the SCOUT metadata
Google Sheet, with columns \code{SPN}, \code{mutant}, \code{type}, \code{CNA_type}, \code{chr},
\code{start}, \code{end}, \code{ref}, \code{alt}, \code{code}, \code{allele}, \code{src_allele}.
Optionally filtered to a specific SPN.
}
\examples{
\dontrun{
get_ground_truth_drivers()
get_ground_truth_drivers("SPN01")
}
}
