% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sheets.R
\name{get_ground_truth_exposures}
\alias{get_ground_truth_exposures}
\title{Get SCOUT ground truth mutational signature exposures}
\usage{
get_ground_truth_exposures(spn = NULL, sample = NULL, type = NULL)
}
\arguments{
\item{spn}{SPN identifier, e.g. \code{"SPN01"}. If \code{NULL}, returns all SPNs.}

\item{sample}{Sample identifier. If \code{NULL}, returns all samples.}

\item{type}{Signature type, e.g. \code{"SBS"} or \code{"ID"}. If \code{NULL}, returns
all types.}
}
\value{
A tibble.
}
\description{
Returns a tibble of ground truth signature exposures from the SCOUT
metadata Google Sheet, with columns \code{SPN}, \code{Sample}, \code{Coverage}, \code{Purity},
\code{Type}, \code{Signature}, and \code{Exposure}.
Optionally filtered by SPN, sample, and/or signature type.
}
\examples{
\dontrun{
get_ground_truth_exposures()
get_ground_truth_exposures("SPN01")
get_ground_truth_exposures("SPN01", type = "SBS")
}
}
