% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_path.R
\name{get_lambda1_max_glasso}
\alias{get_lambda1_max_glasso}
\title{Smallest value of lambda1 which makes cov(X,X) have all zero off-diagonal
entries. The diagonal matrix will have diagonal entries
1/(lambda1_max + x_i^T * x_i)}
\usage{
get_lambda1_max_glasso(X, desc_frac = 0.8, max_iter = 100, use_huge = TRUE)
}
\description{
Smallest value of lambda1 which makes cov(X,X) have all zero off-diagonal
entries. The diagonal matrix will have diagonal entries
1/(lambda1_max + x_i^T * x_i)
}
