% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_path.R
\name{get_lambda1_path}
\alias{get_lambda1_path}
\title{Get the lambda1 path dynamically, from the data. The data is assumed to
be standardized already.}
\usage{
get_lambda1_path(
  X,
  p1,
  nlambda = 100,
  lambda_min_ratio = 0.001,
  desc_frac = 0.8,
  max_iter = 100
)
}
\arguments{
\item{X}{data matrix (n x p)}

\item{p1}{type of regularization in first scout step. Can be NULL or 1.
Ignored if greater than 1}

\item{nlambda}{how many lambda1's do you want to regularize over?}

\item{lambda_min_ratio}{smallest value for lambda1 as a fraction of
lambda1_max.}

\item{desc_frac}{by how much to multiply lambda_max when tuning it in
bisection search for the "ideal" lambda_max}

\item{max_iter}{maximum number of iterations to use to find the lowest
lambda_max in bisection search}
}
\description{
Since scout only uses lambda1 to regularize the XX precision matrix,
this function only needs the X data
}
