% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_path.R
\name{get_lambda2_max_lasso}
\alias{get_lambda2_max_lasso}
\title{get_lambda2_max_lasso}
\usage{
get_lambda2_max_lasso(X, Y, X_cov, desc_frac = 0.8, max_iter = 100)
}
\arguments{
\item{X}{feature matrix (n x p)}

\item{Y}{response vector (n x 1)}

\item{desc_frac}{default = 0.8. fraction of previous lambda_max to multiply
when descending to find smallest lambda_max}

\item{max_iter}{default = 100. maximum number of iterations}

\item{cov_X}{covariance of X data. if using with lasso_one, pass in the
covariance that comes back from lasso_one. otherwise can pass in cov(X)}
}
\description{
computes the lambda2 max in the scout procedure for a fixed
covariance estiamte. the lambda2 max is the smallest value of lambda that
shrinks all coefficients to zero.
}
