% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout_alternating.R
\name{get_xtxs}
\alias{get_xtxs}
\title{Computes covariance matrix estimate using either L1 or L2 penalty}
\usage{
get_xtxs(X, p1, lam1s)
}
\arguments{
\item{X}{n x p data matrix. It is assumed to be standardized}

\item{p1}{numeric value, either 1 or 2, corresponding to L1 and L2
regularization respectively. If p1 = 1, huge::huge.glasso is used to estimate
a sparse covariance matrix. If p1 = 2, scout::gridge2 is used to estimate
a covariance matrix.}

\item{lam1s}{numeric vector of penalty parameters for the regularization}
}
\value{
a list of length length(lam1s) where each entry is a covariance
matrix estimate. List indices correspond to lam1s indices.
}
\description{
Computes covariance matrix estimate using either L1 or L2 penalty
}
