% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_yhat}
\alias{get_yhat}
\title{This predict function is for a single beta, single new_x matrix}
\usage{
get_yhat(betahat, meanx, meany, sdx, sdy, newx)
}
\arguments{
\item{meany}{(maybe robust) mean of y data used to fit the model}

\item{new_x}{X matrix on which to predict}
}
\description{
This predict function is for a single beta, single new_x matrix
}
