% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso_select.R
\name{glasso_select}
\alias{glasso_select}
\title{`glasso_select`}
\usage{
glasso_select(
  X,
  nlambda = 100,
  lambda.min.ratio = 0.1,
  crit = "bic",
  standardize = TRUE,
  scr = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{X}{feature matrix, \eqn{n \times p}}

\item{nlambda}{number of lambdas to optimize over}

\item{lambda.min.ratio}{smallest value of lambda as a fraction of lambda_max}

\item{crit}{criteria to select the optimal lambda. one of "bic" or "loglik"}

\item{scr}{whether to use lossy screening in huge.glasso}

\item{verbose}{whether to let huge.glasso print progress messages}
}
\value{
list of:
* `icov`: matrix - inverse covariance estimate based on best lambda
* `best_lambda`: numeric - best lambda selected based on `crit`
* `lambda`: numeric vector - sequence of lambdas that was used for selection
* `errors`: numeric vector - `crit` values corresponding to each value in
`lambda`
}
\description{
Uses huge::huge.glasso to estimate an inverse covariance matrix given
a data matrix.
}
