% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvglasso.R
\name{glasso_traintest}
\alias{glasso_traintest}
\title{Compute glasso validation error for a particular train/test split
Assume X_train and X_test are NOT already standardized}
\usage{
glasso_traintest(X_train, X_test, lambda, alternateCov, crit, scr, verbose)
}
\arguments{
\item{X_train}{unstandardized train data}

\item{X_test}{unstandardized test data}

\item{lambda}{lambda squence (numeric vector)}
}
\description{
Compute glasso validation error for a particular train/test split
Assume X_train and X_test are NOT already standardized
}
