% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridge_select.R
\name{gridge_select}
\alias{gridge_select}
\title{gridge_select}
\usage{
gridge_select(
  X,
  nlambda = 100,
  lambda.min.ratio = 0.1,
  crit = "bic",
  standardize = TRUE
)
}
\arguments{
\item{X}{feature matrix, \eqn{n \times p}}

\item{nlambda}{number of lambdas to optimize over}

\item{lambda.min.ratio}{smallest value of lambda as a fraction of lambda_max}

\item{crit}{criteria to select the optimal lambda. one of "bic" or "loglik"
#' @return list of:
* `w`: matrix - inverse precision matrix estimate based on best lambda
* `best_lambda`: numeric - best lambda selected based on `crit`
* `lambda`: numeric vector - sequence of lambdas that was used for selection
* `errors`: numeric vector - `crit` values corresponding to each value in
`lambda`}
}
\description{
Uses scout::gridge2 to estimate an inverse *precision* matrix given a
data matrix.
}
