% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso_select.R
\name{huge_glasso_lambda_seq}
\alias{huge_glasso_lambda_seq}
\title{`huge_glasso_lambda_seq`}
\usage{
huge_glasso_lambda_seq(cov_X, nlambda, lambda.min.ratio)
}
\arguments{
\item{cov_X}{covariance matrix estimate}

\item{nlambda}{number of lambdas to return}

\item{lambda.min.ratio}{smallest value of lambda as a fraction of lambda_max}
}
\value{
numeric vector of length nlambda, in decreasing order, of log-spaced
lambda values
}
\description{
Computes the lambda sequence for huge::huge.glasso
}
