% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso_select.R
\name{icov_error}
\alias{icov_error}
\title{Error of inverse covariance matrix estimate}
\usage{
icov_error(icov, cov, n, method = "loglik")
}
\arguments{
\item{icov}{inverse covariance estimate}

\item{cov}{covariance estimate}

\item{n}{number of rows in original data matrix}

\item{method}{one of "loglik", "bic"}
}
\description{
Error of inverse covariance matrix estimate
}
