% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zenodo.R
\name{list_zenodo_files}
\alias{list_zenodo_files}
\title{List files available in a Zenodo record}
\usage{
list_zenodo_files(record_id)
}
\arguments{
\item{record_id}{Zenodo record ID (integer or string).}
}
\value{
A tibble with columns \code{filename}, \code{size}, \code{download_url}.
}
\description{
List files available in a Zenodo record
}
