% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{log_message}
\alias{log_message}
\title{Log messages to a file and print optionally}
\usage{
log_message(message, log_file, verbose = FALSE)
}
\arguments{
\item{message}{Text to write}

\item{log_file}{Path to log file}

\item{verbose}{Logical, whether to print message}
}
\description{
Log messages to a file and print optionally
}
