% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LT.smooth.all}
\alias{LT.smooth.all}
\title{Smoothing function for gene expression}
\usage{
LT.smooth.all(tree_, lognorm__, ka, s = NULL, v = "v3", norm = "B")
}
\description{
Smoothing function for gene expression
}
