% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.scoutobject}
\alias{predict.scoutobject}
\title{predict.scoutobject}
\usage{
\method{predict}{scoutobject}(object, newx, ...)
}
\arguments{
\item{object}{The results of a call to the "scout" function. The
coefficients that are part of this object will be used for
making predictions}

\item{newx}{The new x at which predictions should be made. Can be a
vector of length ncol(x), where x is the data on which scout.obj was
created, or a matrix with ncol(x) columns}

\item{...}{Additional arguments to predict}
}
\value{
`yhat`: If newx was a vector, then a  matrix will be returned,
with dimension length(lam1s)xlength(lam2s) (where lam1s and lam2s
are attributes of scout.obj). The (i,j) element of this matrix will
correspond to tuning parameter values (lam1s[i], lam2s[j]). If newx
is a matrix, then an array of dimension
nrow(newx) x length(lam1s) x length(lam2s) will be returned.
}
\description{
A function to perform prediction, using an x matrix and the output of
the "scout" function.
}
