% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCOUT.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data and inputs for OUwie modeling}
\usage{
prepare_data(
  tree_path,
  metadata_path,
  outpath,
  species_key = NULL,
  quant_traits = NULL,
  name = NULL,
  regimes = "BM1",
  algorithm = "three.point",
  anc_infer = "ape",
  resolve = "dichotomous",
  normalize = FALSE,
  smoothing = NULL,
  floor = 0,
  scale = FALSE,
  blacklist = NULL,
  troubleshoot = FALSE,
  write_smooth = FALSE,
  logfile = NULL
)
}
\arguments{
\item{tree_path}{Path to the phylogenetic tree file in Newick format}

\item{metadata_path}{Path to the metadata CSV file}

\item{outpath}{Output directory path}

\item{species_key}{Column name for species ID in metadata (optional)}

\item{quant_traits}{Vector or file path of quantitative trait names or 'infer'}

\item{name}{Prefix for output files}

\item{regimes}{Vector of regime column names or single regime name}

\item{algorithm}{OUwie algorithm}

\item{anc_infer}{Ancestral state inference method: 'ape', 'castor', or 'skip'}

\item{resolve}{Tree multifurcation resolution method}

\item{normalize}{Logical, whether to log-normalize data}

\item{smoothing}{Optional integer smoothing parameter}

\item{floor}{Numeric floor for zero counts}

\item{scale}{T/F whether to scale tree height to 1.}

\item{blacklist}{Vector of column names to exclude when inferring quantitative traits}

\item{troubleshoot}{Logical, whether to limit genes for troubleshooting}

\item{write_smooth}{Logical, whether to save smoothed counts intermediate file}
}
\value{
List with processed inputs, metadata, gene columns, regimes, and attributes
}
\description{
Prepare data and inputs for OUwie modeling
}
