% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{process_ouwie_single}
\alias{process_ouwie_single}
\title{Process OUwie model results into data frame summarizing parameters and model fit}
\usage{
process_ouwie_single(res, gene, regime)
}
\description{
Process OUwie model results into data frame summarizing parameters and model fit
}
