% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{processes_results}
\alias{processes_results}
\title{Process results to pick winners based on delta AICc thresholds}
\usage{
processes_results(i, d = 2, ac = 0.1)
}
\description{
Process results to pick winners based on delta AICc thresholds
}
