% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sheets.R
\name{read_sheet}
\alias{read_sheet}
\title{Read a public Google Sheet and return a tibble}
\usage{
read_sheet(sheet_id, gid = NULL)
}
\arguments{
\item{sheet_id}{The Google Sheets document ID (the long string in the URL).}

\item{gid}{Optional sheet tab ID (integer). Defaults to the first sheet.}
}
\value{
A tibble.
}
\description{
Read a public Google Sheet and return a tibble
}
\examples{
\dontrun{
read_sheet("1BxiMVs0XRA5nFMdKvBdBZjgmUUqptlbs74OgVE2upms")
}
}
