% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout_alternating.R
\name{rescale_betas}
\alias{rescale_betas}
\title{Rescales beta estimates as in Step 4 of the Scout algo (Witten & Tibshirani)}
\usage{
rescale_betas(X_train, Y_train, betamat)
}
\arguments{
\item{X_train}{standardized training data}

\item{Y_train}{standardized training response}

\item{betamat}{a matrix where each column is a beta_hat estimate for some
regularization parameter}
}
\description{
Rescales beta estimates as in Step 4 of the Scout algo (Witten & Tibshirani)
}
