% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{return_model}
\alias{return_model}
\title{Wrapper function to return the model results}
\usage{
return_model(model_input, meta, regime, candidate, options)
}
\description{
Wrapper function to return the model results
}
