% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_standardize.R
\name{rob_standardize}
\alias{rob_standardize}
\title{Returns list of x, y, meanx, meany, sdx, sdy}
\usage{
rob_standardize(x, y, standardize = TRUE, alternateCov = "default")
}
\arguments{
\item{x}{A matrix, n x p}

\item{y}{A vector or n x 1 matrix}

\item{standardize}{TRUE or FALSE, indicating whether standardization should occur.}

\item{alternateCov}{A string indicating which alternate covariance estimator to use. Currently implemented options are "default" and "cellwise"}
}
\value{
list with x, y, meanx, meany, sdx, and sdy
}
\description{
Returns list of x, y, meanx, meany, sdx, sdy
}
