% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{run_one_ouwies_one_gene}
\alias{run_one_ouwies_one_gene}
\title{Run OUwie model on a single gene and regime}
\usage{
run_one_ouwies_one_gene(phy, meta, regime, gene, model, algorithm, mode, sh)
}
\description{
Run OUwie model on a single gene and regime
}
