% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOUT_EM_utils.R
\name{runSCOUT.batches}
\alias{runSCOUT.batches}
\title{Main runner but in batches, helpful for large trees.}
\usage{
runSCOUT.batches(
  idata,
  lambda1,
  lambda2,
  fixed.root,
  weight_type = "alt",
  algorithm = "three.point",
  tau_prior_mean = 0.2,
  tau_prior_sd = 0.1,
  M_only = FALSE,
  skipTau = FALSE,
  runid = NULL,
  cores = 1,
  batch_size = 50,
  results_dir = "results_tmp",
  logfile = NULL,
  verbose = FALSE
)
}
\description{
Main runner but in batches, helpful for large trees.
}
