% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOUT_EM_utils.R
\name{runSCOUT}
\alias{runSCOUT}
\title{Main runner}
\usage{
runSCOUT(
  idata,
  lambda1,
  lambda2,
  fixed.root,
  weight_type = "alt",
  algorithm = "three.point",
  tau_prior_mean = 0.2,
  tau_prior_sd = 0.1,
  alpha_shape = 2,
  alpha_rate = 1,
  M_only = FALSE,
  skipTau = FALSE,
  scaleHeight = FALSE,
  runid = NULL,
  cores = 1,
  logfile = NULL,
  verbose = FALSE
)
}
\description{
Main runner
}
