\name{scout-internal}
\title{Internal scout functions}
\alias{print.scoutobject}
\alias{print.cvobject}
\description{Internal scout functions}
\usage{
print.scoutobject(x,...)
print.cvobject(x,...)
}
\author{Daniela Witten and Robert Tibshirani}
\details{These functions are not to be called by the user.}
\keyword{internal}




