% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout_stepwise.R
\name{scout_1something_stepwise}
\alias{scout_1something_stepwise}
\title{Scout(1,.) Stepwise}
\usage{
scout_1something_stepwise(
  X,
  Y,
  p2,
  K = 5,
  nlambda1 = 100,
  lambda1.min.ratio = 0.1,
  nlambda2 = 100,
  lambda2.min.ratio = 0.1,
  standardize = TRUE
)
}
\arguments{
\item{X}{n x p data matrix}

\item{Y}{n x 1 response vector}

\item{p2}{regularization type in second step. one of NULL or 1}

\item{K}{number of folds to use in CV, only applicable for p2 = 1}

\item{nlambda1}{number of regularization params to use in glasso step}

\item{lambda1.min.ratio}{smallest value of lambda1 as a fraction of lambda1_max}

\item{nlambda2}{number of regularization params to use in lasso step}

\item{lambda2.min.ratio}{smallest value of lambda2 as a fraction of lambda2_max}
}
\value{
list of
* `g.res`: result of graphical lasso step
* `cv.res`: result of cross-validated lasso step, or NA if p2 = NULL
* `mod`: final scout model using cross-validated parameters
}
\description{
Computes the best inverse covariance estimate using the glasso and uses that
in the second step of the Scout.
}
