% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout_alternating.R
\name{scout_alternating_lasso}
\alias{scout_alternating_lasso}
\title{scout_alternating_lasso}
\usage{
scout_alternating_lasso(
  X_train,
  Y_train,
  X_test,
  Y_test,
  p1,
  nlambda1 = 100,
  nlambda2 = 100,
  lambda1_min_ratio = 0.1,
  lambda2_min_ratio = 0.001,
  tol = 1e-04,
  max_iter = 10,
  rescale = TRUE,
  standardize = TRUE,
  lam1_init = "random"
)
}
\arguments{
\item{X_train}{A matrix of predictors. Rows are observations and columns
are variables}

\item{Y_train}{A vector of outcomes}

\item{X_test}{A matrix of predictors (test set)}

\item{Y_test}{A vector of outcomes (test set)}

\item{p1}{L_p penalty for covariance regularization. must be 1 or 2}

\item{nlambda1}{number of regularization terms to use in covariance
regularization step}

\item{nlambda2}{number of regularization terms to use in coefficient
regularization step}

\item{lambda1_min_ratio}{smallest fraction of lambda1_max to include in
lambda1 sequence}

\item{lambda2_min_ratio}{smallest fraction of lambda2_max to include in
lambda2 sequence}

\item{tol}{convergence tolerance for difference between previous estimated
rmspe and current estimated rmspe. if the difference becomes smaller than
this value, the algorithm stops and returns the current lam1 and lam2
combination}

\item{max_iter}{maximum number of alternating iterations to compute before
stopping. the loop will stop when either the number of iterations exceeds
max_iter, or when the errors converge to have a diff below the tolerance}

\item{rescale}{Should coefficients beta obtained by
covariance-regularized regression be re-scaled by a constant, given
by regressing $y$ onto $x beta$? This is done in Witten and
Tibshirani (2008) and is important for good performance. Default is
TRUE.}

\item{standardize}{whether or not to scale the training X and Y data
before performing estimation}

\item{lam1_init}{one of "random" or "max"}
}
\value{
list with items:
* errors: vector of numeric RMSPE values
* betas: list of vectors of estimated beta_hats
* intercepts: vector of numeric intercept values
* lambda_pairs: list of vectors of length 2, each vector is (lam1, lam2)
* lambda2_paths: list of vectors of lambda2 paths. user usually doesn't need
this but it's provided as a convenience
Incides of each list or vector (other than lambda2_paths) correspond to
the lambda pair at that index in lambda_pairs
}
\description{
implements an alternating search algorithm to efficiently
search over a grid of lambda1 x lambda2. L_p penalty for coefficient
regularization is assumed to be 1 (i.e., p2 = 1)
}
