% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout_stepwise.R
\name{scout_glasso_stepwise}
\alias{scout_glasso_stepwise}
\title{Puts together glasso_select and scout1something}
\usage{
scout_glasso_stepwise(
  X,
  Y,
  p2 = 1,
  K = 10,
  nlambda1 = 100,
  lambda1.min.ratio = 0.1,
  nlambda2 = 100,
  lambda2.min.ratio = 0.01,
  alternateCov = "default"
)
}
\arguments{
\item{X}{n \times p data matrix}

\item{Y}{n \times 1 response vector}

\item{p2}{regularization type in second step. one of NULL or 1}

\item{K}{number of folds to use in CV. only applicable for p2 = 1}

\item{nlambda1}{number of regularization params to use in glasso step}

\item{nlambda2}{number of regularization params to use in lasso step}

\item{alternateCov}{one of "cellwise" or "default"}
}
\value{
list of
* `g.res`: result of graphical lasso step
* `cv.res`: result of cross-validated lasso step, or NA if p2 = NULL
* `betahat`: coefficient estimate#' * `yhat`: predictions on test data
* `mod`: final scout model using cross-validated parameters
}
\description{
Puts together glasso_select and scout1something
}
