% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_lineage}
\alias{simulate_lineage}
\title{Calculate fit metrics and weights from model summaries}
\usage{
simulate_lineage(ngenes, ncells, tree, outdir, out_prefix)
}
\arguments{
\item{ngenes}{Integer describing number of genes to simuate for the tedsim pipeline.}

\item{ncells}{Integer describing number of cells to simulate.}

\item{tree}{Newick string describing cell state tree. Controls how many final states are simulated.}

\item{outdir}{Path to output directory}

\item{out_prefix}{Prefix for saving files.}
}
\value{
List of results. The same set are also saved to outdir.
}
\description{
Calculate fit metrics and weights from model summaries
}
