% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCOUT.R
\name{get_fit_metrics}
\alias{get_fit_metrics}
\title{Calculate fit metrics and weights from model summaries}
\usage{
get_fit_metrics(
  models_summary,
  outpath = NULL,
  prefix = "OUWIE",
  write = FALSE
)
}
\arguments{
\item{models_summary}{Data frame returned from fitModel}

\item{outpath}{Directory to write weighted results CSV}

\item{prefix}{File prefix}

\item{write}{Logical, whether to write weighted results CSV}
}
\value{
Data frame with AICc ranks, delta and weights
}
\description{
Calculate fit metrics and weights from model summaries
}
