% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvglasso.R
\name{glasso_select}
\alias{glasso_select}
\title{Uses huge::huge.glasso to estimate an inverse covariance matrix given
a data matrix.}
\usage{
glasso_select(
  X,
  nlambda = 10,
  lambda.min.ratio = 0.1,
  standardize = TRUE,
  alternateCov = "default",
  crit = "bic",
  scr = F,
  verbose = F
)
}
\arguments{
\item{X}{feature matrix, n \times p}

\item{nlambda}{number of lambdas to optimize over}

\item{lambda.min.ratio}{smallest value of lambda as a fraction of lambda_max}

\item{standardize}{whether or not to standardize the columns of X before
performing glasso}

\item{alternateCov}{either "default" or "cellwise". if "default", passes
cov(X) to huge.glasso. if "cellwise", passes the cellwise robust covariance
estimate of X to huge.glasso.}

\item{crit}{criteria to select the optimal lambda. one of "bic" or "loglik"}

\item{scr}{whether to use lossy screening in huge.glasso}

\item{verbose}{whether to let huge.glasso print progress messages}
}
\value{
list of:
* `icov`: matrix - inverse covariance estimate based on best lambda,
* `best_lambda`: numeric - best lambda selected based on `crit`
* `lambda`: numeric vector - sequence of lambdas that was used for selection
* `errors`: numeric vector - `crit` values corresponding to each value in
`lambda`
}
\description{
Uses huge::huge.glasso to estimate an inverse covariance matrix given
a data matrix.
}
