% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout.R
\name{print.scoutobject}
\alias{print.scoutobject}
\title{Prediction function for covariance-regularized regression, aka the Scout.}
\usage{
\method{print}{scoutobject}(x, ...)
}
\arguments{
\item{...}{Additional arguments to predict}

\item{objet}{The results of a call to the "scout" function. The
coefficients that are part of this object will be used for
making predictions.}

\item{newx}{The new x at which predictions should be made. Can be a
vector of length ncol(x), where x is the data on which scout.obj was
created, or a matrix with ncol(x) columns.}
}
\description{
A function to perform prediction, using an x matrix and the output of
the "scout" function.
}
\details{
predict.scoutobject
}
