% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SCOUT}
\alias{SCOUT}
\title{Wrapper function to run SCOUT}
\usage{
SCOUT(
  counts.file,
  tree.file,
  results_dir,
  regimes,
  species_key = "species",
  blacklist = NULL,
  method = "EM",
  testid = NULL,
  normalize = TRUE,
  scale_tree = FALSE,
  smoothing_k = NULL,
  tau_prior_sd = 0.1,
  tau_prior_mean = 0.2,
  fixed_root = FALSE,
  lambda1 = 0.2,
  lambda2 = 0.2,
  cores = 1,
  logfile = NULL,
  verbose = TRUE
)
}
\arguments{
\item{counts.file}{path to a counts file with a columns for species and relevent columns for evolutionary models being tested if an OUX. Other columns will be genes.}

\item{tree.file}{path to newick file. tips match species names in counts.}

\item{results_dir}{output directory}

\item{regimes}{a list of regimes to test. This include specifying BM1, OU1, and OUx where OUx correspond to columns in the counts file.}

\item{species_key}{if the species column is not called 'species' (rather cellBC for example), indicate the correct name here.}

\item{blacklist}{genes for testing are inferred automatically. if any columns in the counts file are not for testing and not either the species key or model name, list them here.}

\item{method}{one of EM (expectation-maximization), SM (smoothing), MTF (tip fog).}
}
\description{
Wrapper function to run SCOUT
}
