% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_test_data}
\alias{simulate_test_data}
\title{Calculate fit metrics and weights from model summaries}
\usage{
simulate_test_data(
  ngenes,
  ncells,
  tree,
  outdir,
  out_prefix,
  a,
  s,
  t0,
  theta_step,
  nevfs = 20,
  sim_lin = TRUE,
  param_mode = "rank",
  signal_gain = c(NA, NA, 0.3),
  burst_scale = 1,
  scale_s = 10
)
}
\arguments{
\item{ngenes}{Integer describing number of genes to simuate for the tedsim pipeline.}

\item{ncells}{Integer describing number of cells to simulate. Note ncells < ngenes.}

\item{tree}{Newick string describing cell state tree. Controls how many final states are simulated.}

\item{outdir}{Path to output directory}

\item{out_prefix}{Prefix for saving files.}

\item{a}{Integer or list: Alpha to simulate values.}

\item{s}{Integer of list: Sigma to simulate values.}

\item{t0}{Integer: Root theta values.}

\item{theta_step}{Integer: Used to generate optimal values depending on the model.}

\item{nevfs}{Integer: Number of EVFs to generate.}

\item{param_mode}{Character: Either rank or affine. Controls whether OU-trait scale is preseved when moving to counts.}

\item{signal_gain}{List: Default c(NA,NA,0.3). Controls the how much to scale the variance of an OU process before converting to counts.}

\item{burst_scale}{Integer: Default = 1. Controls level of transcriptional bursting.}

\item{scale_s}{Integer: Default = 10. Controls the rate of synthesis.}
}
\value{
List of results. The same set are also saved to outdir.
}
\description{
Calculate fit metrics and weights from model summaries
}
